Rename to state package
