+++ src/main/groovy/io/seqera/wave/service/cleanup/CleanupStore.groovy
@@ -19,10 +19,9 @@
 package io.seqera.wave.service.cleanup
 
 import groovy.transform.CompileStatic
-import io.seqera.wave.memstore.range.AbstractRangeStore
-import io.seqera.wave.memstore.range.impl.RangeProvider
+import io.seqera.wave.store.range.AbstractRangeStore
+import io.seqera.wave.store.range.impl.RangeProvider
 import jakarta.inject.Singleton
-
 /**
  * Implements a timed range to store and retrieve IDs
  * of resources to be cleaned up.
