+++ src/main/groovy/io/seqera/wave/service/mirror/MirrorStateStore.groovy
@@ -22,13 +22,13 @@ import java.time.Duration
 
 import groovy.transform.CompileStatic
 import io.seqera.wave.encoder.MoshiEncodeStrategy
-import io.seqera.wave.service.cache.AbstractCacheStore
-import io.seqera.wave.service.cache.impl.CacheProvider
+import io.seqera.wave.service.state.AbstractCacheStore
+import io.seqera.wave.service.state.impl.CacheProvider
 import jakarta.inject.Inject
 import jakarta.inject.Singleton
 
 /**
- * Implement a {@link io.seqera.wave.service.cache.CacheStore} for {@link MirrorState} objects
+ * Implement a {@link io.seqera.wave.service.state.CacheStore} for {@link MirrorState} objects
  *
  * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
