+++ src/main/groovy/io/seqera/wave/service/state/StateRecord.groovy
@@ -16,14 +16,15 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package io.seqera.wave.service.cache.impl
-
-import io.seqera.wave.service.cache.CacheStore
+package io.seqera.wave.service.state
 
 /**
- * Define an cache interface alias to be used by cache implementation providers
- *
+ * Marker interface for record object that model long running operation state
+ * 
  * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
-interface CacheProvider<K,V> extends CacheStore<K,V> {
+interface StateRecord {
+
+    String getRecordId()
+
 }
