+++ src/main/groovy/io/seqera/wave/store/range/RangeStore.groovy
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package io.seqera.wave.memstore.range
+package io.seqera.wave.store.range
 /**
  * Define the contract for a storage range set similar to Redis `zrange`
  *
