+++ src/main/groovy/io/seqera/wave/tower/auth/JwtAuthStore.groovy
@@ -24,8 +24,8 @@ import java.time.Instant
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
 import io.seqera.wave.encoder.MoshiEncodeStrategy
-import io.seqera.wave.service.cache.AbstractCacheStore
-import io.seqera.wave.service.cache.impl.CacheProvider
+import io.seqera.wave.service.state.AbstractCacheStore
+import io.seqera.wave.service.state.impl.CacheProvider
 import jakarta.inject.Inject
 import jakarta.inject.Singleton
 /**
